Extract shared validation methods from facilities and nearby controllers
