+++ modules/va_facilities/app/controllers/va_facilities/v0/facilities_controller.rb
@@ -6,18 +6,19 @@ require_dependency 'va_facilities/application_controller'
 require_dependency 'va_facilities/pagination_headers'
 require_dependency 'va_facilities/geo_serializer'
 require_dependency 'va_facilities/csv_serializer'
+require_dependency 'va_facilities/param_validators'
 
 module VaFacilities
   module V0
     class FacilitiesController < ApplicationController
       include ActionController::MimeResponds
       include VaFacilities::PaginationHeaders
+      include VaFacilities::ParamValidators
       skip_before_action(:authenticate)
       before_action :set_default_format
       before_action :validate_params, only: [:index]
 
       REQUIRE_ONE_PARAM = %i[bbox long lat zip ids state].freeze
-      TYPE_SERVICE_ERR = 'Filtering by services is not allowed unless a facility type is specified'
       MISSING_PARAMS_ERR =
         'Must supply lat and long, bounding box, zip code, or ids parameter to query facilities data.'
 
@@ -105,36 +106,6 @@ module VaFacilities
         end
       rescue ArgumentError
         raise Common::Exceptions::InvalidFieldValue.new('bbox', params[:bbox])
-      end
-
-      def validate_no_services_without_type
-        if params[:type].nil? && params[:services].present?
-          raise Common::Exceptions::ParameterMissing.new('type', detail: TYPE_SERVICE_ERR)
-        end
-      end
-
-      def validate_type_and_services_known
-        raise Common::Exceptions::InvalidFieldValue.new('type', params[:type]) unless
-          BaseFacility::TYPES.include?(params[:type])
-        unknown = params[:services].to_a - BaseFacility::SERVICE_WHITELIST[params[:type]]
-        raise Common::Exceptions::InvalidFieldValue.new('services', unknown) unless unknown.empty?
-      end
-
-      def validate_state_code
-        if params[:state] && STATE_CODES.exclude?(params[:state])
-          raise Common::Exceptions::InvalidFieldValue.new('state', params[:state])
-        end
-      end
-
-      def validate_zip
-        if params[:zip]
-          raise Common::Exceptions::InvalidFieldValue.new('zip', params[:zip]) unless
-            params[:zip].match?(/\A\d{5}(-\d{4})?\z/)
-          zip_plus0 = params[:zip][0...5]
-          requested_zip = ZCTA.select { |area| area[0] == zip_plus0 }
-          raise Common::Exceptions::InvalidFieldValue.new('zip', params[:zip]) unless
-            requested_zip.any?
-        end
       end
 
       def valid_location_query?
