+++ modules/va_facilities/app/controllers/va_facilities/v1/nearby_controller.rb
@@ -6,18 +6,19 @@ require_dependency 'va_facilities/application_controller'
 require_dependency 'va_facilities/pagination_headers'
 require_dependency 'va_facilities/geo_serializer_v1'
 require_dependency 'va_facilities/csv_serializer'
+require_dependency 'va_facilities/param_validators'
 
 module VaFacilities
   module V1
     class NearbyController < ApplicationController
       include ActionController::MimeResponds
       include VaFacilities::PaginationHeaders
+      include VaFacilities::ParamValidators
       skip_before_action(:authenticate)
       before_action :set_default_format
       before_action :validate_params, only: [:index]
 
       REQUIRED_PARAMS = %i[street_address city state zip].freeze
-      TYPE_SERVICE_ERR = 'Filtering by services is not allowed unless a facility type is specified'
       MISSING_PARAMS_ERR =
         'Must supply street_address, city, state, and zip to query nearby facilities.'
 
@@ -71,40 +72,10 @@ module VaFacilities
         end
       end
 
-      def validate_state_code
-        if params[:state] && STATE_CODES.exclude?(params[:state])
-          raise Common::Exceptions::InvalidFieldValue.new('state', params[:state])
-        end
-      end
-
-      def validate_zip
-        if params[:zip]
-          raise Common::Exceptions::InvalidFieldValue.new('zip', params[:zip]) unless
-            params[:zip].match?(/\A\d{5}(-\d{4})?\z/)
-          zip_plus0 = params[:zip][0...5]
-          requested_zip = ZCTA.select { |area| area[0] == zip_plus0 }
-          raise Common::Exceptions::InvalidFieldValue.new('zip', params[:zip]) unless
-            requested_zip.any?
-        end
-      end
-
       def validate_drive_time
         Integer(params[:drive_time]) if params[:drive_time]
       rescue ArgumentError
         raise Common::Exceptions::InvalidFieldValue.new('drive_time', params[:drive_time])
-      end
-
-      def validate_no_services_without_type
-        if params[:type].nil? && params[:services].present?
-          raise Common::Exceptions::ParameterMissing.new('type', detail: TYPE_SERVICE_ERR)
-        end
-      end
-
-      def validate_type_and_services_known
-        raise Common::Exceptions::InvalidFieldValue.new('type', params[:type]) unless
-          BaseFacility::TYPES.include?(params[:type])
-        unknown = params[:services].to_a - BaseFacility::SERVICE_WHITELIST[params[:type]]
-        raise Common::Exceptions::InvalidFieldValue.new('services', unknown) unless unknown.empty?
       end
 
       def metadata(resource)
